refactor quota evaluation to cleanly abstract the quota access
